AyenNdTask/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==3.0.3
django-allauth==0.41.0
django-cors-headers==3.2.1
django-crispy-forms==1.9.0
-e git://github.com/kmmbvnr/django-fsm.git@2508065fa94232fe0cdac151582851645f28a8b1#egg=django_fsm
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1
idna==2.9
inflection==0.4.0
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.3
PyJWT==1.7.1
pyparsing==2.4.7
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
